Improved code style of User.jsx
